backend2/requirements.txt
fastapi==0.120.0
uvicorn==0.38.0
python-multipart==0.0.20
requests==2.32.5
python-dotenv==1.1.1
Pillow==12.0.0
